Fix scrutinizer warnings
