+++ app/Http/Controllers/Controller.php
@@ -17,7 +17,6 @@
 
 namespace App\Http\Controllers;
 
-use App\Helpers\SendsAlerts;
 use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 use Illuminate\Foundation\Bus\DispatchesJobs;
 use Illuminate\Foundation\Validation\ValidatesRequests;
