+++ app/Jobs/UpdateServer.php
@@ -6,14 +6,20 @@ use App\Enums\HostStatus;
 use App\Events\HostKeysUpdated;
 use App\Models\Host;
 use App\Services\SFTP\SFTPPusher;
+use Illuminate\Bus\Queueable;
 use Illuminate\Contracts\Queue\ShouldQueue;
 use Illuminate\Foundation\Bus\Dispatchable;
+use Illuminate\Queue\InteractsWithQueue;
+use Illuminate\Queue\SerializesModels;
 use Illuminate\Support\Facades\Storage;
 use phpseclib\Crypt\RSA;
 
 class UpdateServer implements ShouldQueue
 {
     use Dispatchable;
+    use InteractsWithQueue;
+    use Queueable;
+    use SerializesModels;
 
     protected Host       $host;
     protected SFTPPusher $pusher;
