+++ app/Models/Host.php
@@ -74,7 +74,8 @@ class Host extends Model implements Searchable
 
     public function getPortAttribute(string $value): int
     {
-        return $value ?? (int) setting()->get('ssh_port');
+        $defaultPort = setting()->get('ssh_port');
+        return $value ?? (int) $defaultPort;
     }
 
     public function scopeNotInSync(Builder $query): Builder
@@ -123,7 +124,7 @@ class Host extends Model implements Searchable
                             break;
                         case 'allow':
                             $content = explode(' ', $key->public, 3);
-                            $content[2] = $key->username.'@ssham';
+                            $content[2] = $key->username . '@ssham';
                             $sshKeys[$key->username] = join(' ', $content);
                             break;
                         default:
@@ -132,7 +133,7 @@ class Host extends Model implements Searchable
                 }
             }
         }
-        if (! is_null($bastionSSHPublicKey)) {
+        if (!is_null($bastionSSHPublicKey)) {
             $sshKeys[] = $bastionSSHPublicKey;
         }
