+++ tests/Feature/Models/HostTest.php
@@ -219,4 +219,19 @@ class HostTest extends TestCase
         $got = $host->getSSHKeysForHost($bastionHostKey->public);
         $this->assertCount(2, $got);
     }
+
+
+    /** @test */
+    public function getFullHostname_return_the_full_hostname(): void
+    {
+        /** @var Host $host */
+        $host = Host::factory()->makeOne([
+            'username' => 'root',
+            'hostname' => 'server1.domain.local',
+            'port' => 12345,
+        ]);
+        $want = 'root@server1.domain.local:12345';
+
+        $this->assertEquals($want, $host->full_hostname);
+    }
 }
